feat: add redirections. Use new index page. Add new links to header
