--- docs/configuration-interactive-cli.md
@@ -1,6 +1,0 @@
----
-id: configuration-interactive-cli
-title: 301 - Moved permanently
----
-
-This page has been moved to [plugins-inquirer-cli](plugins-inquirer-cli)
